move helper class into its own header file
