+++ searchcore/src/vespa/searchcore/proton/matching/match_tools.h
@@ -2,14 +2,15 @@
 
 #pragma once
 
-#include "queryenvironment.h"
-#include "isearchcontext.h"
-#include "query.h"
-#include "viewresolver.h"
-#include "querylimiter.h"
-#include "match_phase_limiter.h"
+#include "field_id_to_name_mapper.h"
 #include "handlerecorder.h"
+#include "isearchcontext.h"
+#include "match_phase_limiter.h"
+#include "query.h"
+#include "queryenvironment.h"
+#include "querylimiter.h"
 #include "requestcontext.h"
+#include "viewresolver.h"
 #include <vespa/searchcommon/attribute/i_attribute_functor.h>
 #include <vespa/searchlib/queryeval/blueprint.h>
 #include <vespa/searchlib/common/idocumentmetastore.h>
@@ -29,28 +30,6 @@ namespace search::fef {
 }
 
 namespace proton::matching {
-
-class FieldIdToNameMapper {
-public:
-    const std::string & lookup(uint32_t fieldId) const {
-        static const std::string lookupFailed;
-        if (auto fieldInfo = _indexEnv.getField(fieldId)) {
-            return fieldInfo->name();
-        }
-        return lookupFailed;
-    }
-    uint32_t lookup(const std::string& fieldName) const {
-        if (auto fieldInfo = _indexEnv.getFieldByName(fieldName)) {
-            return fieldInfo->id();
-        }
-        return -1;
-    }
-    FieldIdToNameMapper(const search::fef::IIndexEnvironment &indexEnv)
-      : _indexEnv(indexEnv)
-    {}
-private:
-    const search::fef::IIndexEnvironment &_indexEnv;
-};
 
 class MatchTools
 {
